Migrated all distinct functions to their own modules, with defs, and loaded them in botcommands
