+++ slack-irc/lib/twitch.js
@@ -1,5 +1,6 @@
 var http = require('http');
 var https = require('https');
+var logger = require('winston');
 var apikeys = require('./apikeys');
 
 twitchOnlineStreams =[];
@@ -123,4 +124,4 @@ module.exports = {
 	"commands": {
 		"twitch" : twitchOnlineCheck
 	}
-}
+};
